ClipService/requirements.txt
flask
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
Pillow